Added template for Unit Tests
* Update README.md

* trying to get tests up and running

* trying to get tests up and running v2

* trying to get tests up and running v4

* trying to get tests up and running v5

* trying to get tests up and running v6

* Fix config script

* Add pytest to requirements.txt

* Code adjusted to be unit tested

* adding required helper class

* adding required helper class

* adding extra test

---------
