+++ .env.template
@@ -1,4 +1,4 @@
 export LLAMA_MODEL_PATH=""
 export OPENAI_API_KEY=""
 export AI21_API_KEY=""
-export LLM="AI21"
+export LLM="CHATGPT"
