+++ README.md
@@ -21,8 +21,6 @@ An Overview of Project 30 -
 
 # ibm-sweng
 
-# chinaza/RAG-chromadb
-
 ## Setup
 
 Reccomended CPU prerequisite setup.
@@ -43,7 +41,7 @@ Download the "Datafiniti" CSV file from [here](https://data.world/datafiniti/con
 Place it in the project root driectory.
 The program will for a csv with the name `Datafiniti_Amazon_Consumer_Reviews_of_Amazon_Products.csv` in the directory in which it is invoked.
 
-You need to modify `.env.template`.
+You need to modify `.env.template`. (no longer necessary, modify config.py instead)
 If you are using LLaMa set the `LLM` environment variable to `LLAMA` and make sure that the `LLAMA_MODEL_PATH` environment variable points to the `gguf` files corresponding to the model that you want to use.
 If you are using chatgpt, make sure you set the `LLM` environment variable to `CHATGPT` and set `OPENAI_API_KEY` to your OpenAI API key.
 If you are using AI21 llm, make sure you set the `LLM` environment variable to `AI21` and set `AI21_API_KEY` to your AI21 API key.
@@ -51,6 +49,9 @@ If you are using AI21 llm, make sure you set the `LLM` environment variable to `
 ## Running
 
 ```
-. .env.template
+. .env.template (no longer necessary)
 streamlit run ChatbotConcept.py
+
+Please run the commands using bash, rather than powershell as it might not work using powershell (windows devices).
+
 ```
