+++ requirements.txt
@@ -7,3 +7,4 @@ sentence_transformers
 llama-cpp-python
 chromadb
 langchain_openai
+pytest
