+++ test_main.py
@@ -1,4 +1,20 @@
-import unittest
+#UNIT TESTS
 
-def test_placeholder():
-    pass
+#Don't remove these imports
+import ChatbotHelper
+import config
+
+#TODO CREATE TESTS
+def test_get_db():
+    ChatbotHelper.get_db()
+    assert 2 + 2 == 4
+
+def test_get_llm():
+    assert 2 + 2 == 4
+
+def test_get_memory():
+    assert 2 + 2 == 4
+
+def test_get_options():
+    assert 2 + 2 == 4
+
